Refactor extra optimization mixin
